use unsigned integer type for video->aspect attribute

+++ src/player.h
@@ -147,7 +147,7 @@ typedef struct mrl_properties_video_s {
   uint32_t bitrate;
   uint32_t width;
   uint32_t height;
-  float aspect;
+  uint32_t aspect;
   uint32_t channels;
   uint32_t streams;
   float framerate;
